test(congressional): split behavioral render into the components lane (fix CI)
The combined test put a Svelte 5 render() in CI's main lane, which can't mount()
(lifecycle_function_unavailable) — CI runs only vitest.config.ts (test:coverage), which
EXCLUDES render tests; the components lane (mount) is local-only by design.

Split into the two lanes:
- congressional-delivery-tier.test.ts        — STRUCTURAL single-source lock, runs in CI
  (constant declared once, all 3 sites import it, no local literal, no hardcoded tier-4,
   each endpoint gates on it).
- congressional-delivery-gate.behavior.test.ts — BEHAVIORAL render (constituent at the
  floor proceeds, below is gated), components lane; added to vitest.config.ts exclude
  alongside VerificationGate-auto-dismiss et al.

Structural: 4 pass (main lane). Behavioral: 2 pass (components lane). svelte-check 0.
